move regex functions to net_util
